moved HD key derivation code into secp256k1.ts
