+++ README.md
@@ -3,7 +3,7 @@
 # Decentralized Web Node (DWN) SDK
 
 Code Coverage
-![Statements](https://img.shields.io/badge/statements-93.83%25-brightgreen.svg?style=flat) ![Branches](https://img.shields.io/badge/branches-92.4%25-brightgreen.svg?style=flat) ![Functions](https://img.shields.io/badge/functions-91.01%25-brightgreen.svg?style=flat) ![Lines](https://img.shields.io/badge/lines-93.83%25-brightgreen.svg?style=flat)
+![Statements](https://img.shields.io/badge/statements-93.82%25-brightgreen.svg?style=flat) ![Branches](https://img.shields.io/badge/branches-92.39%25-brightgreen.svg?style=flat) ![Functions](https://img.shields.io/badge/functions-91.01%25-brightgreen.svg?style=flat) ![Lines](https://img.shields.io/badge/lines-93.82%25-brightgreen.svg?style=flat)
 
 ## Introduction
